+++ news.md
@@ -1,4 +1,4 @@
-# tidySEM 0.2.11
+# tidySEM 0.2.11.1
 
 * pmc() deprecates pmc_srmr() function
 * Add srmr() function
@@ -6,6 +6,9 @@
 * Export mx_deviances()
 * Fix bug in get_edges() and get_nodes() for multigroup lavaan models with user-
   defined quantities (`:=`), issue #109.
+* Add `use_labels` parameter to `table_results()` for lavaan and OpenMx models,
+  allowing variable names to be replaced with labels from a dictionary. 
+  Closes #56.
 
 # tidySEM 0.2.10